+++ src/saihubot-addon-ethgas.js
@@ -4,7 +4,6 @@ function int(num) {
   return parseInt(num, 10);
 }
 
-// Addon
 export const fetchGas = (fetch, estimator, callback) =>
   fetch(estimator.api).then(response => response.json()).then(json => {
     const data = estimator.processGasData(json);
@@ -49,3 +48,20 @@ export const GAS_ESTIMATOR = {
     }),
   }
 }
+
+// Addon
+/**
+ * fetch Eth estimate Gas Price
+ */
+export const addonFetchGas = {
+  name: 'fetchGas',
+  requirements: {
+    addons: ['fetch'],
+  },
+  action: (robot) => (estimator, callback) =>
+    fetchGas(robot.addons.fetch, estimator, callback),
+};
+
+const addons = [addonFetchGas];
+
+export {addons};